+++ backend/canisters/community/impl/src/queries/c2c_can_issue_access_token_for_channel.rs
@@ -3,7 +3,8 @@ use crate::read_state;
 use crate::RuntimeState;
 use canister_api_macros::query_msgpack;
 use community_canister::c2c_can_issue_access_token_for_channel::*;
-use types::AccessTokenType;
+use group_chat_core::{GroupChatCore, GroupMemberInternal};
+use types::{AccessTokenType, VideoCallType};
 
 #[query_msgpack(guard = "caller_is_local_user_index")]
 fn c2c_can_issue_access_token_for_channel(args: Args) -> Response {
@@ -20,9 +21,30 @@ fn c2c_can_issue_access_token_for_channel_impl(args: Args, state: &RuntimeState)
     };
 
     match args.access_type {
-        AccessTokenType::StartVideoCall => {
-            args.is_diamond && member.role.is_permitted(channel.chat.permissions.start_video_call)
+        AccessTokenType::StartVideoCall => can_start_video_call(
+            member,
+            args.is_diamond,
+            state.data.is_public,
+            VideoCallType::Default,
+            &channel.chat,
+        ),
+        AccessTokenType::StartVideoCallV2(vc) => {
+            can_start_video_call(member, args.is_diamond, state.data.is_public, vc.call_type, &channel.chat)
         }
         AccessTokenType::JoinVideoCall | AccessTokenType::MarkVideoCallAsEnded => true,
     }
+}
+
+fn can_start_video_call(
+    member: &GroupMemberInternal,
+    is_diamond: bool,
+    is_public_community: bool,
+    call_type: VideoCallType,
+    chat: &GroupChatCore,
+) -> bool {
+    if !is_diamond || !member.role.is_permitted(chat.permissions.start_video_call) {
+        return false;
+    }
+
+    !is_public_community || !chat.is_public.value || matches!(call_type, VideoCallType::Broadcast)
 }