+++ backend/canisters/user/impl/src/queries/c2c_can_issue_access_token.rs
@@ -16,7 +16,7 @@ fn c2c_can_issue_access_token_impl(args: Args, state: &RuntimeState) -> bool {
     }
 
     match args.access_type {
-        AccessTokenType::StartVideoCall => args.is_diamond,
+        AccessTokenType::StartVideoCall | AccessTokenType::StartVideoCallV2(_) => args.is_diamond,
         AccessTokenType::JoinVideoCall | AccessTokenType::MarkVideoCallAsEnded => true,
     }
 }